requirements.txt
flask==3.0.3
chromadb==0.5.3
python-dotenv==1.0.1
requests==2.32.3
pytest==8.3.2
gunicorn==22.0.0